template fixes: User follower count

+++ app/models.py
@@ -97,10 +97,18 @@ class User(UserMixin, db.Model):
     followed = db.relationship('Follow', foreign_keys=[Follow.follower_id], backref=db.backref('follower', lazy='joined'), lazy='dynamic', cascade='all, delete-orphan')
     followers = db.relationship('Follow', foreign_keys=[Follow.followed_id], backref=db.backref('followed', lazy='joined'), lazy='dynamic', cascade='all, delete-orphan')
 
+    @staticmethod
+    def add_self_follows():
+        for user in User.query.all():
+            if not user.is_following(user):
+                user.follow(user)
+                db.session.add(user)
+                db.session.commit()
+
 
     def __init__(self, **kwargs):
         super(User, self).__init__(**kwargs)
-
+        self.follow(self)
         if self.role is None:
             if self.email == current_app.config['BLOG_TWO_ADMIN']:
                 self.role = Role.query.filter_by(name='Administrator').first()
@@ -109,6 +117,7 @@ class User(UserMixin, db.Model):
 
         if self.email is not None and self.avatar_hash is None:
             self.avatar_hash = self.gravatar_hash()
+
 
 
     def __repr__(self):
@@ -237,15 +246,18 @@ class User(UserMixin, db.Model):
         return hashlib.md5(self.email.lower().encode('utf-8')).hexdigest()
 
 
-    def follow(self):
-        '''
-        Inserts Follow ins
-        '''
+    def follow(self, user):
+        if user.id is None:
+            return False
+
         if not self.is_following(user):
             f = Follow(follower=self, followed=user)
             db.session.add(f)
 
-    def unfollow(self):
+    def unfollow(self, user):
+        if user.id is None:
+            return False
+
         f = self.followed.filter_by(followed_id=user.id).first()
         if f:
             db.session.delete(f)
@@ -259,11 +271,12 @@ class User(UserMixin, db.Model):
         if user.id is None:
             return False
         return self.followers.filter_by(follower_id=user.id).first() is not None
-        
+
     @property
     def followed_posts(self):
         return Post.query.join(Follow, Follow.followed_id == Post.author_id)\
             .filter(Follow.follower_id == self.id)
+
 
 class AnonymousUser(AnonymousUserMixin):
     def can(self, permissions):
